requirements.txt
groq>=0.9.0
pydantic>=2.0.0
pypdf>=4.0.0
reportlab>=4.0.0
faker>=24.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
